Add subpages to table order and collect order

+++ src/main_website/php_file/minusCart.php
@@ -5,9 +5,10 @@ require_once "../../../src/class/Database.php";
 
 if (isset($_GET['itemID'])) {
     $itemID = $_GET['itemID'];
+    $userID = $_SESSION['usersID'];
 
     $result = Database::query("SELECT * FROM item WHERE itemID ='$itemID'");
-    $resultCart = Database::query("SELECT * FROM cart WHERE itemID = '$itemID'");
+    $resultCart = Database::query("SELECT * FROM cart WHERE itemID = '$itemID' AND usersID ='$userID'");
 
     $rowCart = mysqli_fetch_array($resultCart);
     $row = mysqli_fetch_array($result);
@@ -16,9 +17,10 @@ if (isset($_GET['itemID'])) {
     $cost = $row['price']*$quantity;
 }
 if ($quantity>0){
-    Database::query("UPDATE cart SET quantity = '$quantity', totalPrice = '$cost' WHERE itemID='$itemID'");
+    Database::query("UPDATE cart SET quantity = '$quantity', totalPrice = '$cost' 
+            WHERE itemID='$itemID' AND usersID = '$userID'");
 }else{
-    Database::query("DELETE FROM cart WHERE itemID = '$itemID'");
+    Database::query("DELETE FROM cart WHERE itemID = '$itemID' AND usersID = '$userID'");
 }
 
 Database::disconnect();
